chore: add getAttributionClassValueByLabel
