+++ src/v2/Apps/Artwork/Components/ArtworkSidebar/__tests__/CreateArtworkAlertSection.jest.tsx
@@ -1,9 +1,6 @@
 import { graphql } from "react-relay"
 import { setupTestWrapperTL } from "v2/DevTools/setupTestWrapper"
-import {
-  CreateArtworkAlertSectionFragmentContainer,
-  getAttributionClassIdByLabel,
-} from "../CreateArtworkAlertSection"
+import { CreateArtworkAlertSectionFragmentContainer } from "../CreateArtworkAlertSection"
 import { CreateArtworkAlertSection_Test_Query } from "v2/__generated__/CreateArtworkAlertSection_Test_Query.graphql"
 import { fireEvent, screen } from "@testing-library/react"
 import { useTracking } from "react-tracking"
@@ -81,50 +78,6 @@ describe("CreateArtworkAlertSection", () => {
       context_page_owner_type: "artwork",
       context_page_owner_id: "artwork-id",
       context_page_owner_slug: "artwork-slug",
-    })
-  })
-})
-
-describe("getAttributionClassIdByLabel", () => {
-  it("should return `unique`", () => {
-    const result = getAttributionClassIdByLabel("Unique")
-    expect(result).toEqual(["unique"])
-  })
-
-  it("should return `limited edition`", () => {
-    const result = getAttributionClassIdByLabel("Limited Edition")
-    expect(result).toEqual(["limited edition"])
-  })
-
-  it("should return `open edition`", () => {
-    const result = getAttributionClassIdByLabel("Open Edition")
-    expect(result).toEqual(["open edition"])
-  })
-
-  it("should return `unknown edition`", () => {
-    const result = getAttributionClassIdByLabel("Unknown Edition")
-    expect(result).toEqual(["unknown edition"])
-  })
-
-  describe("Different cases", () => {
-    it("should return correct value when all letters are lowercase", () => {
-      const result = getAttributionClassIdByLabel("unknown edition")
-      expect(result).toEqual(["unknown edition"])
-    })
-
-    it("should return correct value when all letters are uppercase", () => {
-      const result = getAttributionClassIdByLabel("UNKNOWN EDITION")
-      expect(result).toEqual(["unknown edition"])
-    })
-
-    it("should return correct value when all letters are different case", () => {
-      expect(getAttributionClassIdByLabel("Unknown Edition")).toEqual([
-        "unknown edition",
-      ])
-
-      expect(getAttributionClassIdByLabel("UnKnOwN EdItIoN")).toEqual([
-        "unknown edition",
-      ])
     })
   })
 })
